fix: authorization for users

+++ src/controllers/userPlaylist/playlist.controller.ts
@@ -15,6 +15,11 @@ const getAllPlaylists = asyncHandler(
 const getPlaylistByUserId = asyncHandler(
   async (req: Request, res: Response, next: NextFunction) => {
     const { id } = req.params;
+
+    if (id !== req.user.userId.toString().toString()) {
+      throw new ApiError("You cannot view someone else playlist", 403);
+    }
+
     const playlists = await query(
       `SELECT * FROM playlists WHERE user_id = ${id}`
     );
@@ -34,12 +39,17 @@ const getPlaylistByUserId = asyncHandler(
 const getPlaylistDetail = asyncHandler(
   async (req: Request, res: Response, next: NextFunction) => {
     const { id } = req.params;
+
     const playlistQuery = await query(`SELECT playlists.*, 
 ARRAY_AGG(json_build_object('id', songs.id, 'name', songs.name, 'duration_ms', songs.duration_ms, 'album', songs.album, 'artists', songs.artists, 'preview_url', songs.preview_url)) AS tracks 
 FROM playlists 
 LEFT JOIN songs ON playlists.id = songs.playlist_id 
 WHERE playlists.id = ${id} 
 GROUP BY playlists.id;`);
+
+    if (playlistQuery.user_id.toString() !== req.user.userId.toString()) {
+      throw new ApiError("You cannot view someone else playlist", 403);
+    }
 
     const playlist = {
       ...playlistQuery,
@@ -58,6 +68,10 @@ GROUP BY playlists.id;`);
 
 const createPlaylist = asyncHandler(
   async (req: Request, res: Response, next: NextFunction) => {
+    if (req.body.userId !== req.user.userId.toString()) {
+      throw new ApiError("You can create playlist for yourself only", 403);
+    }
+
     await query(
       `INSERT INTO playlists (name, type, description, user_id) VALUES ('${req.body.name}', 'playlist', '${req.body.description}', '${req.body.userId}')`
     );
@@ -77,6 +91,16 @@ const addSongToPlaylist = asyncHandler(
     );
 
     if (!song) {
+      const playlist = await query(`SELECT * FROM playlists WHERE id = ${id}`);
+
+      if (!playlist) {
+        throw new ApiError("Playlist is not available", 404);
+      }
+
+      if (playlist.user_id.toString() !== req.user.userId.toString()) {
+        throw new ApiError("You cannot add song in someone else playlist", 403);
+      }
+
       const artists = req.body.artists.map((artist: any) =>
         JSON.stringify(artist)
       );
@@ -107,9 +131,15 @@ const deleteSongFromPlaylist = asyncHandler(
   async (req: Request, res: Response, next: NextFunction) => {
     const { id } = req.params;
 
-    const song = await query(`SELECT * FROM songs WHERE id = '${id}'`);
+    const song = await query(
+      `SELECT s.id AS song_id, p.user_id as user_id FROM songs s LEFT JOIN playlists p ON p.id = s.playlist_id WHERE s.id = '${id}'`
+    );
 
     if (song) {
+      if (song.user_id.toString() !== req.user.userId.toString()) {
+        throw new ApiError("You cannot delete song someone else playlist", 403);
+      }
+
       await query(`DELETE FROM songs WHERE id = '${id}'`);
 
       res.sendStatus(204);
@@ -125,5 +155,5 @@ export {
   getPlaylistDetail,
   createPlaylist,
   addSongToPlaylist,
-  deleteSongFromPlaylist
+  deleteSongFromPlaylist,
 };
